+++ node/resolvers/items.ts
@@ -49,7 +49,7 @@ export const adjustItems = (
 
     return {
       ...item,
-      imageUrl: fixImageUrl(item.imageUrl)!,
+      imageUrl: fixImageUrl(item.imageUrl, platform)!,
       name: product.productName,
       skuSpecifications: getVariations(item.id, product.items),
     }